+++ bindings/c++/include/uhppoted.hpp
@@ -61,6 +61,7 @@ typedef struct card {
     std::string from;
     std::string to;
     uint8_t doors[4];
+    uint32_t PIN;
 } card;
 
 typedef struct time_profile {
@@ -124,7 +125,7 @@ class uhppoted {
     int get_cards(uint32_t id);
     card get_card(uint32_t id, uint32_t card_number);
     card get_card_by_index(uint32_t id, uint32_t index);
-    void put_card(uint32_t id, uint32_t card_number, std::string from, std::string to, uint8_t doors[4]);
+    void put_card(uint32_t id, uint32_t card_number, std::string from, std::string to, uint8_t doors[4], uint32_t PIN);
     void delete_card(uint32_t id, uint32_t card_number);
     void delete_cards(uint32_t id);