+++ examples/ccl/examples.lisp
@@ -139,12 +139,14 @@
          (from        "2022-01-01")
          (to          "2022-12-31")
          (doors       (make-array 4 :initial-contents '(0 1 31 75)))
-         (ok          (exec #'(lambda (u) (uhppoted-put-card u device-id card-number from to doors)))))
+         (PIN         7531)
+         (ok          (exec #'(lambda (u) (uhppoted-put-card u device-id card-number from to doors PIN)))))
     (when ok)
       (display "put-card" device-id (list "card"  card-number
                                           "from"  from
                                           "to"    to
-                                          "doors" (format nil "~{~a ~}" (coerce doors 'list))))))
+                                          "doors" (format nil "~{~a ~}" (coerce doors 'list))
+                                          "PIN"   PIN))))
 
 
 (defun delete-card (args) ""